% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_authorities.R
\name{get_authorities}
\alias{get_authorities}
\title{Get local authorities}
\usage{
get_authorities(country = NULL)
}
\arguments{
\item{country}{The country that the authority is located within, as a
character string: "England", "Wales", "Scotland" or "Northern Ireland".}
}
\value{
A tibble of authorities and their code, the number of
establishments they have, and for England or all authorities, their region.
}
\description{
Fetches the list of authorities in the FHRS
}
\details{
`get_authorities` makes an API call that returns the authorities that
submit data to the FHRS scheme, and their ID, which can be used to filter
other API calls. It takes an optional country parameter.
}
\examples{
# Get all authorities in the FHRS
get_authorities()

# Get all authorities in the FHRS in Scotland
get_authorities("Scotland")

}
