% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_business_types.R
\name{get_business_types}
\alias{get_business_types}
\title{Get business types}
\usage{
get_business_types()
}
\value{
A tibble of business types and their codes
}
\description{
Fetches the list of business types in the FHRS
}
\details{
`get_business_types` makes an API call that returns the business types that
can be found in the FHRS, and their code, which can be used to filter other
API calls. It takes no parameters.
}
\examples{
get_business_types()

}
