% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_establishments.R
\name{get_establishments}
\alias{get_establishments}
\title{Get establishments}
\usage{
get_establishments(name = NULL, la = NULL, type = NULL,
  condensed = TRUE)
}
\arguments{
\item{name}{A string to search for within the business name}

\item{la}{The ID of the required local authority (as returned from `get_authorities()`)}

\item{type}{The code of the required business type (as returned from `get_business_types()`)}

\item{condensed}{By default the function returns a selection of variables: to return
everything in the API call, set this to FALSE}
}
\value{
A tibble containing all businesses that meet the specified criteria
}
\description{
Fetches a list of establishments
}
\details{
`get_establishments` makes an API call that returns all establishments, with
the ability to filter by name, local authority and business type.
}
\examples{
\dontrun{
library(dplyr)
# A list of all restaurants in Aberdeen
# Find the code for Aberdeen
get_authorities() \%>\%
  filter(grepl("Aberdeen", name))
# Find the code for restaurants
get_business_types()
# Perform establishment search
get_establishments(la = 197, type = 1)

# A list of all businesses containing the words "pret a manger" in Birmingham
# Find out the code for Birmingham
get_authorities() \%>\%
  filter(grepl("Birmingham", name))
# Perform establishment search
get_establishments(name = "pret a manger", la = 374)
}


}
